pyproject.toml
[tool.poetry]
name = "past1000"
version = "0.1.0"
description = "Processing data from CMIP6 for 'past 1000' experiment."
authors = ["SongshGeo <songshgeo@gmail.com>"]
readme = "README.md"
include = [
    'config/variables.yaml',
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
loguru = "*"
hydra-core = "~1.3"
xarray = ">=2023"
matplotlib = ">=3.9.2"
netcdf4 = ">=1.7"
rioxarray = ">=0.17.0"
geopandas = "~1"
pint-xarray = ">=0.4"
nc-time-axis = ">=1.4"
matplotkit = ">=0.1.0"
# 使用 poetry run python -m pip install climate-indices 安装
# 不知道为什么，使用 poetry 安装的 climate-indices 版本无法运行
# 参考 https://climate-indices.readthedocs.io/en/latest/
# climate-indices = ">=2.0.1"
tqdm = "*"
cartopy = ">=0.24.1"
openpyxl = "*"
xclim = "^0.53.2"
cf-xarray = "*"
mksci-font = ">=0.1.3"
multi-ruptures = ">=0.1.3"
seaborn = "^0.13.2"
arviz = "^0.20.0"
pymc = "^5.20.1"
GeoDSKit = {path = "../GeoDSKit"}
fitter = "^1.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
allure-pytest = "^2.13.5"
pytest-clarity = "^1.0.1"
pytest-sugar = "^1.0.0"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
